fix(visa-engine): break import cycle via _types leaf + list() enum iteration (CodeQL)
Clears 12 CodeQL alerts on #2718: 8x py/unsafe-cyclic-import (models/ast/
fact_registry mutual imports) + 4x py/non-iterable-in-for-loop (str-Enum
iteration in tests). Extracts the shared leaf types (UnknownReason,
Applicant/DerivedFactPath, FactPath, KnownFact, UnknownFact, FactSnapshot)
into _types.py (stdlib-only leaf) and re-exports them from fact_registry so
every existing import path stays valid — no public API change, no consumer
ripple. models/ast now import from _types, not fact_registry; the only edge
back is fact_registry's TYPE_CHECKING import of ApplicantFacts (single
directed edge, no cycle). 223 tests green, import chain + re-exports verified.
